+++ src/Framework/Config.php
@@ -4,6 +4,9 @@ namespace Parable\Framework;
 
 class Config
 {
+    /** @var string */
+    protected $mainConfigClass = \Config\App::class;
+
     /** @var \Parable\Filesystem\Path */
     protected $path;
 
@@ -14,6 +17,21 @@ class Config
         \Parable\Filesystem\Path $path
     ) {
         $this->path = $path;
+    }
+
+    /**
+     * @param string $className
+     *
+     * @return $this
+     * @throws Exception
+     */
+    public function setMainConfigClassName($className)
+    {
+        if (!class_exists($className)) {
+            throw new \Parable\Framework\Exception("Main Config class '{$className}' does not exist.");
+        }
+        $this->mainConfigClass = $className;
+        return $this;
     }
 
     /**
@@ -37,7 +55,7 @@ class Config
      */
     public function get($key)
     {
-        $config = $this->getConfig();
+        $config = $this->getAll();
         if (strpos($key, '.') !== false) {
             return $this->getNested($config, explode('.', $key));
         }
@@ -50,7 +68,7 @@ class Config
     /**
      * @return array
      */
-    public function getConfig()
+    public function getAll()
     {
         return $this->config;
     }
@@ -60,84 +78,30 @@ class Config
      */
     public function load()
     {
-        $configPath = $this->path->getDir("app/Config");
-
-        if (!file_exists($configPath)) {
-            // Quietly fail, since we don't require the path to exist
+        try {
+            $this->addConfig(\Parable\DI\Container::get($this->mainConfigClass));
+        } catch (\Exception $e) {
+            // We don't throw here since the file isn't required to exist, but we do stop here
             return $this;
         }
 
-        $dirIterator = new \RecursiveDirectoryIterator(
-            $configPath,
-            \RecursiveDirectoryIterator::SKIP_DOTS
-        );
-        $iteratorIterator = new \RecursiveIteratorIterator($dirIterator);
-
-        $configClasses = [];
-        foreach ($iteratorIterator as $file) {
-            /** @var \SplFileInfo $file */
-            $className = 'Config\\' . str_replace('.php', '', $file->getFilename());
-
-            /** @var \Parable\Framework\Config\Base $configClass */
-            try {
-                $configClass = \Parable\DI\Container::get($className);
-
-                // @codeCoverageIgnoreStart
-                $configClasses[] = $configClass;
-                // @codeCoverageIgnoreEnd
-            } catch (\Exception $e) {
-                // Just continue, this isn't a major problem.
+        if ($this->get("parable.configs")) {
+            foreach ($this->get("parable.configs") as $configClass) {
+                $this->addConfig(\Parable\DI\Container::get($configClass));
             }
         }
-
-        $this->addConfigs($configClasses);
 
         return $this;
     }
 
     /**
-     * Add a Config file to the stack, ignoring sort order since the moment this is called decides where it fits.
-     *
-     * @param \Parable\Framework\Config\Base $config
+     * @param \Parable\Framework\Interfaces\Config $config
      *
      * @return $this
      */
-    public function addConfig(\Parable\Framework\Config\Base $config)
+    public function addConfig(\Parable\Framework\Interfaces\Config $config)
     {
-        $this->config = array_merge($this->config, $config->getValues());
-        return $this;
-    }
-
-    /**
-     * @param \Parable\Framework\Config\Base[] $configClasses
-     *
-     * @return $this
-     * @throws \Parable\Framework\Exception
-     */
-    public function addConfigs(array $configClasses)
-    {
-        $configs = [];
-        foreach ($configClasses as $configClass) {
-            $configClassName = get_class($configClass);
-            if (!$configClass instanceof \Parable\Framework\Config\Base) {
-                throw new \Parable\Framework\Exception(
-                    "'{$configClassName}' does not implement '\Parable\Framework\Config\Base'"
-                );
-            }
-            if (isset($configs[$configClass->getSortOrder()])) {
-                throw new \Parable\Framework\Exception("Sort order duplication by '{$configClassName}'");
-            }
-            $configs[$configClass->getSortOrder()] = $configClass;
-        }
-
-        // Sort the array by key so the sortOrder is reflected in the actual order
-        ksort($configs);
-
-        // Now get all the values from the config classes
-        foreach ($configs as $configClass) {
-            $this->config = array_merge($this->config, $configClass->getValues());
-        }
-
+        $this->config = array_merge($this->config, $config->get());
         return $this;
     }
 }