+++ src/Framework/Interfaces/Config.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Parable\Framework\Interfaces;
+
+interface Config extends \Parable\Framework\Interfaces\Gettable
+{
+}