+++ src/Framework/Interfaces/Gettable.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Parable\Framework\Interfaces;
+
+interface Gettable
+{
+    /**
+     * Returns a (multidimensional) array with values
+     *
+     * @return array
+     */
+    public function get();
+}