+++ src/Framework/Interfaces/Routing.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Parable\Framework\Interfaces;
+
+interface Routing extends \Parable\Framework\Interfaces\Gettable
+{
+}