+++ tests/Components/Framework/ConfigTest.php
@@ -13,23 +13,27 @@ class ConfigTest extends \Parable\Tests\Components\Framework\Base
     /** @var \Parable\Tests\TestClasses\Config2 */
     protected $config2;
 
-    /** @var \Parable\Tests\TestClasses\Config3DuplicateSortOrder */
-    protected $config3DuplicateSortOrder;
-
     protected function setUp()
     {
         parent::setUp();
 
         $this->config = \Parable\DI\Container::get(\Parable\Framework\Config::class);
 
-        $this->config1                   = new \Parable\Tests\TestClasses\Config1();
-        $this->config2                   = new \Parable\Tests\TestClasses\Config2();
-        $this->config3DuplicateSortOrder = new \Parable\Tests\TestClasses\Config3DuplicateSortOrder();
+        $this->config1 = new \Parable\Tests\TestClasses\Config1();
+        $this->config2 = new \Parable\Tests\TestClasses\Config2();
     }
 
     public function testLoadFailsSilently()
     {
         $this->assertSame($this->config, $this->config->load());
+    }
+
+    public function testSetMainConfigClassThrowsExceptionIfNotExist()
+    {
+        $this->expectException(\Parable\Framework\Exception::class);
+        $this->expectExceptionMessage("Main Config class 'Nope.' does not exist.");
+
+        $this->config->setMainConfigClassName("Nope.");
     }
 
     public function testGetConfig()
@@ -73,42 +77,5 @@ class ConfigTest extends \Parable\Tests\Components\Framework\Base
 
         $this->config->addConfig($this->config2);
         $this->assertSame("secondary value", $this->config->get('setting'));
-    }
-
-    public function testGetConfigSetsDataBasedOnSortOrder()
-    {
-        $this->config->addConfigs([
-            $this->config1,
-            $this->config2,
-        ]);
-
-        $this->config->addConfig($this->config2);
-        $this->assertSame("secondary value", $this->config->get('setting'));
-    }
-
-    public function testAddConfigsThrowsExceptionOnDuplicateSortOrder()
-    {
-        $this->expectException(\Parable\Framework\Exception::class);
-        $this->expectExceptionMessage(
-            "Sort order duplication by 'Parable\Tests\TestClasses\Config3DuplicateSortOrder'"
-        );
-
-        $this->config->addConfigs([
-            $this->config1,
-            $this->config3DuplicateSortOrder,
-        ]);
-    }
-
-    public function testAddConfigsThrowsExceptionOnWrongClass()
-    {
-        $this->expectException(\Parable\Framework\Exception::class);
-        $this->expectExceptionMessage(
-            "'stdClass' does not implement '\Parable\Framework\Config\Base'"
-        );
-
-        $this->config->addConfigs([
-            $this->config1,
-            new \stdClass(),
-        ]);
     }
 }