+++ tests/TestClasses/Autoloader/Basic.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Autoloader;
+
+class Basic
+{
+    public $value = "new";
+
+    public function index()
+    {
+    }
+}