+++ tests/TestClasses/Command/Test.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Parable\Tests\TestClasses\Command;
+
+class Test extends \Parable\Console\Command
+{
+}