+++ tests/TestClasses/Config/TestBrokenRouting.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Parable\Tests\TestClasses\Config;
+
+class TestBrokenRouting implements
+    \Parable\Framework\Interfaces\Config
+{
+    public function get()
+    {
+        return [
+            "parable" => [
+                "routes" => [
+                    \Routing\Wrong::class,
+                ],
+            ],
+        ];
+    }
+}