+++ tests/TestClasses/Init/Test.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Parable\Tests\TestClasses\Init;
+
+class Test
+{
+}