+++ tests/TestClasses/Routing/App.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Routing;
+
+class App implements
+    \Parable\Framework\Interfaces\Routing
+{
+    public function get()
+    {
+        return [];
+    }
+}