+++ tests/TestClasses/Routing/Wrong.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Routing;
+
+class Wrong implements
+    \Parable\Framework\Interfaces\Gettable
+{
+    public function get()
+    {
+        return [];
+    }
+}